Update formatLocalized() with Buddhist Era
Sorry for my newbie to GitHub.

I'm using Laravel 4 and Carbon is recommended package to use with this Framework to manipulate date and time. it's very good package, but i need Buddhist Era in my client's website.

Because of strftime() can convert day and month to locale language, but no luck on year.

so i've modified your formatLocalized(), it's not a gorgeous way but it work for me.

i'm not sure about IntlDateFormatter Class. i've check on php.net, this might help a lot on Localization date and time but it's quite new(php 5.3.0).

PS. sorry for my English.