Add source for isLongYear()
This source helps explain why we've used the 28th of December to determine if a year is a long year. It is the only day of the year that *always* falls on week 52 or 53 of the year, and not year 1 of the following year. All the other dates in the last week of December will occasionally fall on week 51 or week 1.