Fixed issue #883: reverted pull-request #548 about issue #88
The realisation of initial issue #88 is wrong. The problem does not relies to `local` / `not local` timezone. Php handles date changes similar for both cases. So in fact changes was made 'fixes' non-local dates but local dates works same as before.
And it 'fixes' the problem and not actually fixes it because in fact the PHP handling of DST changing in non-UTC dates is quite right. If developer don't want to adjusting timestamp to DST change he should use UTC timezone.

The unit tests added to assure default php behavior.

Next commits reverted by hand because conflicts in automatic revert.

* Commit 2cc69881  Fix CS	07.12.2015 13:04	Lucas Michot
* Commit 5394301d  Fix issue 88	07.12.2015 12:55	Lucas Michot