Add Diff translations for Dutch
There is still one string that I could  not translate,  since that does
not exist in Dutch.

If anyone else knows the word for "before yesterday" in English, please
leave a comment on the PR.

"Voorgisteren" would be the closest I can imagine, but its not correct.

Closing briannesbitt/Carbon#1272

Signed-off-by: MegaXLR <admin@megaxlr.net>