Print original params in Carbon\Traits\Units::addUnit error message
Since PHP 7.0, functions inspecting arguments, like func_get_args(), no longer report the original value as passed to a parameter, but will instead provide the current value.
The parameter "$unit" was changed in previous lines. So the exception message contains the changed value.
This commit preserves the original arguments in variable to be eventually used in the error message.