Fix incorrect documentation for CarbonInterval
Documentation says `CarbonInterval::days(3.5)` returns `3.0hrs` in Carbon 2, but it actually returns `3.0 days`. I assume the latter part of this example is where the mistake lies rather than the code itself, as `CarbonInterval::days(3.5)` is used in the next code block as an example, and it makes sense that these should match.